Rearrange UI display components in components folder
