+++ src/components/BooksUI/Books.jsx
@@ -1,0 +1,11 @@
+import Bookinfo from './Bookinfo';
+import Bookinput from './Bookinput';
+
+const Books = () => (
+  <div className="booksContainer">
+    <Bookinfo />
+    <Bookinput />
+  </div>
+);
+
+export default Books;
