--- src/redux/books/books.js
@@ -1,11 +1,0 @@
-import Bookinfo from './Bookinfo';
-import Bookinput from './Bookinput';
-
-const Books = () => (
-  <div className="booksContainer">
-    <Bookinfo />
-    <Bookinput />
-  </div>
-);
-
-export default Books;
